api/requirements/base.txt
pytz==2019.3  # https://github.com/stub42/pytz
redis==3.3.11  # https://github.com/antirez/redis

# Django
# ------------------------------------------------------------------------------
django==3.0.4  # pyup: < 3.0  # https://www.djangoproject.com/
django-environ==0.4.5  # https://github.com/joke2k/django-environ
django-model-utils==4.0.0  # https://github.com/jazzband/django-model-utils
django-redis==4.11.0  # https://github.com/niwinz/django-redis


# Django REST Framework
djangorestframework==3.11.0  # https://github.com/encode/django-rest-framework
PyYAML==5.2  # https://github.com/yaml/pyyaml
coreapi==2.3.3  # https://github.com/core-api/python-client

# Logging
django-request-logging==0.7.0  # https://github.com/Rhumbix/django-request-logging

# CORS
django-cors-headers==3.2.1